Refactor secret manager plugins and settings
- Moved AWS, Azure, and GCP secret manager implementations to dedicated backend modules.
- Updated the import paths in tests and other modules to reflect the new structure.
- Simplified the settings management by consolidating Pydantic models for AWS, Azure, and GCP.
- Introduced a new registry for backend plugins to facilitate dynamic loading.
- Implemented an in-memory caching mechanism for secret retrieval.
- Enhanced the SecretKeeper class to support various backend configurations.
- Updated tests to ensure compatibility with the new plugin architecture and backend loading mechanism.
