+++ genaikeys/plugins/__init__.py
@@ -1,0 +1,4 @@
+from .base import SecretManagerPlugin
+from .registry import available_backends, load_backend
+
+__all__ = ["SecretManagerPlugin", "available_backends", "load_backend"]
